% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinsu875_Lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class for computing Linear Regression using OLS}
\arguments{
\item{formula}{A Formula.}

\item{data}{A Data frame.}
}
\value{
nothing
}
\description{
This class has a varied methods.
}
\details{
Package Description
}
\section{Fields}{

\describe{
\item{\code{reg_Coef}}{to find regression coefficients}

\item{\code{fit_Val}}{for the fitted values}

\item{\code{res}}{for residuals}

\item{\code{dof}}{for degrees of freedom}

\item{\code{Sigma_square}}{for residual variance}

\item{\code{Var_Beta}}{for Variance of regression coefficients}

\item{\code{t_Beta}}{for t-values for each coefficient}

\item{\code{pvalue}}{for p-values for each coefficient}

\item{\code{parse}}{to parse the input data}

\item{\code{stand_res}}{for standardised residuals for plot2}

\item{\code{variance}}{for variance values}
}}

\section{Methods}{

\describe{
\item{\code{print()}}{Prints information about model}
}}

